+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Enh #51: Add `HasHttpEquiv` trait and `httpEquiv()` method to manage `http-equiv` attribute for HTML elements (@terabytesoftw)
 - Enh #52: Add `HasName` trait and `name()` method to manage `name` attribute for HTML elements (@terabytesoftw)
 - Bug #53: Update documentation for various HTML attributes and their usage in elements (@terabytesoftw)
+- Enh #54: Add `HasDownload` trait and `download()` method to manage `download` attribute for HTML elements (@terabytesoftw)
 
 ## 0.5.2 January 29, 2026